+++ tests/cli/test_init_pandas.py
@@ -7,7 +7,6 @@ from click.testing import CliRunner
 from great_expectations import DataContext
 from great_expectations.cli import cli
 from great_expectations.data_context.util import file_relative_path
-from great_expectations.exceptions import ConfigNotFoundError
 from great_expectations.util import gen_directory_tree_str
 from tests.cli.test_cli import yaml
 
@@ -262,58 +261,3 @@ great_expectations/
         validations/
 """
     )
-
-
-def test_cli_config_not_found(tmp_path_factory):
-    tmp_dir = str(tmp_path_factory.mktemp("test_cli_config_not_found"))
-    curdir = os.path.abspath(os.getcwd())
-    try:
-        os.chdir(tmp_dir)
-        runner = CliRunner()
-        error_message = ConfigNotFoundError().message
-
-        # datasource list
-        result = runner.invoke(cli, ["datasource", "list", "-d", "./"])
-        assert error_message in result.output
-        result = runner.invoke(cli, ["datasource", "list"])
-        assert error_message in result.output
-
-        # datasource new
-        result = runner.invoke(cli, ["datasource", "new", "-d", "./"])
-        assert error_message in result.output
-        result = runner.invoke(cli, ["datasource", "new"])
-        assert error_message in result.output
-
-        # datasource profile
-        result = runner.invoke(cli, ["datasource", "profile", "-d", "./", "--no-view"])
-        assert error_message in result.output
-        result = runner.invoke(cli, ["datasource", "profile", "--no-view"])
-        assert error_message in result.output
-
-        # docs build
-        result = runner.invoke(cli, ["docs", "build", "-d", "./", "--no-view"])
-        assert error_message in result.output
-        result = runner.invoke(cli, ["docs", "build", "--no-view"])
-        assert error_message in result.output
-
-        # project check-config
-        result = runner.invoke(cli, ["project", "check-config", "-d", "./"])
-        assert error_message in result.output
-        result = runner.invoke(cli, ["project", "check-config"])
-        assert error_message in result.output
-
-        # suite new
-        result = runner.invoke(cli, ["suite", "new", "-d", "./"])
-        assert error_message in result.output
-        result = runner.invoke(cli, ["suite", "new"])
-        assert error_message in result.output
-
-        # suite edit
-        result = runner.invoke(cli, ["suite", "edit", "FAKE", "FAKE", "-d", "./"])
-        assert error_message in result.output
-        result = runner.invoke(cli, ["suite", "edit", "FAKE", "FAKE"])
-        assert error_message in result.output
-    except:
-        raise
-    finally:
-        os.chdir(curdir)
